refactor: RegionCountMarkersContainer 으로 개명
[#811]
