--- frontend/src/components/google-maps/marker/CityClusteredMarkersContainer/index.ts
@@ -1,3 +1,0 @@
-import CityClusteredMarkersContainer from './CityClusteredMarkersContainer';
-
-export default CityClusteredMarkersContainer;
