+++ frontend/src/components/google-maps/marker/MarkersContainers.tsx
@@ -4,13 +4,13 @@ import { useExternalValue } from '@utils/external-state';
 
 import { zoomStore } from '@stores/google-maps/zoomStore';
 
-import CityClusteredMarkersContainer from './CityClusteredMarkersContainer';
+import RegionCountMarkersContainer from './RegionCountMarkersContainer';
 import ServerClusteredMarkersContainer from './ServerClusteredMarkersContainer';
 import StationMarkersContainer from './StationMarkersContainer';
 
 const MemoizedStationMarkersContainer = memo(StationMarkersContainer);
 const MemoizedServerClusteredMarkersContainer = memo(ServerClusteredMarkersContainer);
-const MemoizedCityClusteredMarkersContainer = memo(CityClusteredMarkersContainer);
+const MemoizedRegionCountMarkersContainer = memo(RegionCountMarkersContainer);
 
 const MarkersContainers = () => {
   const markerMode = useExternalValue(zoomStore);
@@ -20,7 +20,7 @@ const MarkersContainers = () => {
       {markerMode.state === 'high' && <MemoizedStationMarkersContainer />}
       {/* 이 아래는 앞으로 추가될 기능을 미리 대응하는 컴포넌트 */}
       {markerMode.state === 'middle' && <MemoizedServerClusteredMarkersContainer />}
-      {markerMode.state === 'low' && <MemoizedCityClusteredMarkersContainer />}
+      {markerMode.state === 'low' && <MemoizedRegionCountMarkersContainer />}
     </>
   );
 };
