+++ frontend/src/components/google-maps/marker/RegionCountMarkersContainer/RegionCountMarker.style.ts
