+++ frontend/src/components/google-maps/marker/RegionCountMarkersContainer/index.ts
@@ -1,0 +1,3 @@
+import RegionCountMarkersContainer from './RegionCountMarkersContainer';
+
+export default RegionCountMarkersContainer;
